Working on updater. Should download only new files

+++ src/updater.py
@@ -57,6 +57,15 @@ class Updater():
     def __init__(self, force_update):
         self.web_version = ""
         self.web_files = []
+        self.local_version = info.CNCHI_VERSION
+        
+        # Get local info (local update.info)
+        with open("/usr/share/cnchi/update.info", "r") as local_update_info:
+            response = local_update_info.read()
+            if len(response) > 0:
+                updateInfo = json.loads(response)
+                #self.local_version = updateInfo['version']
+                self.local_files = updateInfo['files']
 
         # Download update.info (contains info of all Cnchi's files)
         update_info_url = _url_prefix + "update.info"
@@ -72,11 +81,12 @@ class Updater():
                 self.force = force_update
 
     def is_web_version_newer(self):
+        """ Returns true if the Internet version of Cnchi is newer than the local one """
         if self.force:
              return True
 
         # Version is always: x.y.z
-        cur_ver = info.CNCHI_VERSION.split(".")
+        cur_ver = self.local_version.split(".")
         web_ver = self.web_version.split(".")
 
         cur = [int(cur_ver[0]),int(cur_ver[1]),int(cur_ver[2])]
@@ -93,6 +103,14 @@ class Updater():
 
         return False
 
+    def should_update_remote_file(self, remote_name, remote_md5):
+        """ Checks if remote file is different from the local one (just compares md5)"""
+        for local_file in self.local_files:
+            if local_file['name'] == remote_name:
+                if local_file['md5'] != remote_md5:
+                    return True
+        return False
+
     def update(self):
         ''' Check if a new version is available and
             update all files only if necessary (or forced) '''
@@ -100,14 +118,17 @@ class Updater():
             logging.info(_("New version found. Updating installer..."))
             num_files = len(self.web_files)
             i = 1
-            for fil in self.web_files:
-                name = fil['name']
-                md5 = fil['md5']
-                print("Downloading %s (%d/%d)" % (name, i, num_files))
-                if self.download(name, md5) is False:
-                    # download has failed
-                    logging.error(_("Download of %s has failed, update will stop"), name)
-                    return False
+            for web_file in self.web_files:
+                name = web_file['name']
+                md5 = web_file['md5']
+                if self.should_update_local_file(name, md5):
+                    print("Downloading %s (%d/%d)" % (name, i, num_files))
+                    if self.download(name, md5) is False:
+                        # download has failed
+                        logging.error(_("Download of %s has failed, update will stop"), name)
+                        return False
+                else:
+                    print("Skipping %s as has not changed" % name)
                 i += 1
             # replace old files with the new ones
             self.replace_old_with_new_versions()
@@ -156,10 +177,9 @@ class Updater():
             new_name = os.path.join(_base_dir, name + "." + self.web_version.replace(".", "_"))
             cur_name = os.path.join(_base_dir, name)
 
-            if os.path.exists(name):
+            # Check that there is a new remote file before deleting the local one
+            if os.path.exists(new_name) and os.path.exists(name):
                 # Remove old file
                 os.remove(name)
-
-            if os.path.exists(new_name):
                 # Rename new download file (removes trailing version in filename)
                 os.rename(new_name, cur_name)
